==> code/05_robustness.R <==

# code/05_robustness.R ----------------------------------------------------------
suppressPackageStartupMessages({
  library(tidyverse); library(fixest); library(MASS); library(modelsummary); library(broom); library(here)
})
source(here::here("R","utils.R")); log_msg("05_robustness: start")

d <- readRDS(here::here("output","intermediate","analytic_harmonized.rds"))
cluster_var <- cluster_picker(d)
ctrl_candidates <- c("HHSize","RESPFemale","HHH_Sex","HHH_Age",
                     "HHH_Education","HHUrbRur","asset_index","earners")
ctrls <- names(d)[str_detect(names(d),
                             paste0("^(", paste(ctrl_candidates, collapse="|"), ")$"))]
ctrl_fml <- if (length(ctrls) > 0) paste(ctrls, collapse = " + ") else "1"

mods <- list()

# Enumerator FE if available
pick_one <- function(cands, nms) {
  for (c in cands) {
    c_regex <- gsub("~", ".*", c)
    hits <- nms[str_detect(nms, regex(c_regex, ignore_case = TRUE))]
    if (length(hits) > 0) return(hits[1])
  }
  NA_character_
}
enum_var <- pick_one(c("EnuName","EnuName_Di~y","EnuName_Diay","EnuName_Diry"), names(d))
if (!is.na(enum_var)) {
  outcomes <- names(d)[str_detect(names(d),
                                  paste0("^(", paste(c("FCS","rCSI","FES","LCS_crisem"),
                                                     collapse="|"), ")$"))]
  for (y in outcomes) {
    f <- as.formula(paste0(y, " ~ Remote + ", ctrl_fml, " | country + ", enum_var))
    mods[[paste0("enumFE_", y)]] <- feols(f, data = d, cluster = cluster_var)
  }
}

# Winsorized FES
if ("FES" %in% names(d)) {
  d <- d %>% mutate(FES_w = winsorize(FES))
  f <- as.formula(paste0("FES_w ~ Remote + ", ctrl_fml, " | country"))
  mods[["FES_w"]] <- feols(f, data = d, cluster = cluster_var)
}

# Categorical: LCS_crisem via logit
if ("LCS_crisem" %in% names(d)) {
  f <- as.formula(paste0("LCS_crisem ~ Remote + ", ctrl_fml, " + factor(country)"))
  mods[["LCS_logit"]] <- glm(f, data = d, family = binomial())
}

tex <- modelsummary(mods, output = "latex",
                    gof_map = c("n","r.squared","rmse","aic","bic"))
write_tex(tex, here::here("output","tables","TableA_robustness.tex"))
tidied <- purrr::map_df(names(mods), ~ broom::tidy(mods[[.x]], conf.int = TRUE) %>% mutate(model = .x))
readr::write_csv(tidied, here::here("output","tables","TableA_robustness.csv"))
log_msg("05_robustness: complete")
